requirements.txt
accelerate==1.4.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
clip==0.2.0
contourpy==1.3.1
cycler==0.12.1
diffusers==0.32.2
einops==0.8.1
filelock==3.13.1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.6.1
ftfy==6.3.1
huggingface-hub==0.25.2
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.4
kiwisolver==1.4.8
kornia==0.8.0
kornia_rs==0.1.8
lightning-utilities==0.12.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
open_clip_torch==2.31.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.3.0
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scipy==1.15.2
six==1.17.0
sympy==1.13.1
timm==1.0.15
tokenizers==0.13.3
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchmetrics==1.6.1
torchvision==0.21.0+cu118
tqdm==4.67.1
transformers==4.28.1
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
yarl==1.18.3
zipp==3.21.0
